+++ backend/src/api/schedules/schedules.controller.ts
@@ -4,24 +4,17 @@ import {
 	Param,
 	ParseUUIDPipe,
 	Patch,
-	UploadedFiles,
 	UseGuards,
 	UseInterceptors,
 } from '@nestjs/common';
-import { FilesInterceptor } from '@nestjs/platform-express';
 import { ApiTags } from '@nestjs/swagger';
 
-import {
-	PatchScheduleTitleSwagger,
-	PatchScheduleUploadThumbnailImageSwagger,
-} from '@/common/decorators/swagger/swagger-schedule.decorator';
-import { BadRequestServiceException } from '@/common/exception/service.exception';
+import { PatchScheduleTitleSwagger } from '@/common/decorators/swagger/swagger-schedule.decorator';
 import { AccessTokenGuard } from '@/common/guards/accessToken.guard';
 import { LoggingInterceptor } from '@/common/interceptors/logging.interceptor';
 import { TimeoutInterceptor } from '@/common/interceptors/timeout.interceptor';
 import { parseUUIDPipeMessage } from '@/common/pipe-message/parse-uuid-pipe-message';
 import { ScheduleUpdateTitleDto } from '@/models/dto/schedule/req/schedule-update-title.dto';
-import { createScheduleThumbnailImageMulterOptions } from '@/utils/upload-media';
 
 import { SchedulesService } from './schedules.service';
 
@@ -55,40 +48,5 @@ export class SchedulesController {
 			scheduleId,
 			dto.scheduleName,
 		);
-	}
-
-	/**
-	 * @summary 특정 스케줄 여행스케줄 썸네일 변경
-	 *
-	 * @tag schedules
-	 * @param {string} scheduleId - 스케줄 아이디
-	 * @param {Express.MulterS3.File} files - 업로드 파일
-	 * @author YangGwangSeong <soaw83@gmail.com>
-	 * @returns string[]
-	 */
-	@PatchScheduleUploadThumbnailImageSwagger()
-	@Patch(':scheduleId/uploads/thumbnail-image')
-	@UseInterceptors(
-		FilesInterceptor('files', 1, createScheduleThumbnailImageMulterOptions()),
-	)
-	async PatchScheduleUploadThumbnailImage(
-		@UploadedFiles() files: Express.MulterS3.File[],
-		@Param(
-			'scheduleId',
-			new ParseUUIDPipe({ exceptionFactory: parseUUIDPipeMessage }),
-		)
-		scheduleId: string,
-	) {
-		if (!files?.length) {
-			throw BadRequestServiceException(`파일이 없습니다.`);
-		}
-		const locations = files.map(({ location }) => location);
-
-		await this.schedulesService.updateScheduleThumbnail(
-			scheduleId,
-			locations[0],
-		);
-
-		return locations;
 	}
 }
