+++ backend/src/common/decorators/swagger/swagger-media.decorator.ts
@@ -3,11 +3,15 @@ import {
 	ApiBadRequestResponse,
 	ApiBody,
 	ApiConsumes,
+	ApiNotFoundResponse,
 	ApiOperation,
 	ApiResponse,
 } from '@nestjs/swagger';
 
-import { ERROR_FILE_NOT_FOUND } from '@/constants/business-error';
+import {
+	ERROR_FILE_NOT_FOUND,
+	ERROR_SCHEDULE_NOT_FOUND,
+} from '@/constants/business-error';
 
 export const PostUploadProfileSwagger = () => {
 	return applyDecorators(
@@ -89,6 +93,34 @@ export const PostUploadFeedMediasSwagger = () => {
 		}),
 		ApiBadRequestResponse({
 			description: ERROR_FILE_NOT_FOUND,
+		}),
+	);
+};
+
+export const PatchScheduleUploadThumbnailImageSwagger = () => {
+	return applyDecorators(
+		ApiConsumes('multipart/form-data'),
+		ApiBody({
+			schema: {
+				type: 'object',
+				properties: {
+					files: {
+						type: 'string',
+						format: 'binary',
+					},
+				},
+			},
+		}),
+		ApiOperation({
+			summary: '특정 스케줄 썸네일 수정하기',
+		}),
+		ApiResponse({
+			description: '스케줄 썸네일 수정하기 성공',
+			isArray: true,
+			type: String,
+		}),
+		ApiNotFoundResponse({
+			description: `1. ${ERROR_SCHEDULE_NOT_FOUND} \n2. ${ERROR_FILE_NOT_FOUND}`,
 		}),
 	);
 };
